# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Agri-PoD Inventory & Distribution System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List campaigns
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [Draft, Submitted, Approved, Active, Completed, Cancelled]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Campaigns
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedCampaigns"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCampaignBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    put:
      operationId: updateCampaign
      tags: [campaigns]
      summary: Update campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCampaignBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}/approve:
    post:
      operationId: approveCampaign
      tags: [campaigns]
      summary: Approve campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}/submit:
    post:
      operationId: submitCampaign
      tags: [campaigns]
      summary: Submit campaign for approval
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
...
  /allocations:
    get:
      operationId: listAllocations
      tags: [allocations]
      summary: List farmer allocations
      parameters:
        - name: campaignId
          in: query
          schema:
            type: integer
        - name: farmerId
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Allocations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedAllocations"
    post:
      operationId: createAllocation
      tags: [allocations]
      summary: Allocate farmer to campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAllocationBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Allocation"
  /allocations/bulk:
    post:
      operationId: bulkAllocateFarmers
      tags: [allocations]
      summary: Bulk allocate farmers to campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkAllocationBody"
      responses:
        "201":
          description: Bulk created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkAllocationResult"

  # VEHICLES
...
  /dispatch:
    get:
      operationId: listDispatches
      tags: [dispatch]
      summary: List vehicle dispatches
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: campaignId
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Dispatches
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedDispatches"
    post:
      operationId: createDispatch
      tags: [dispatch]
      summary: Create vehicle dispatch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDispatchBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dispatch"
  /dispatch/{id}:
    get:
      operationId: getDispatch
      tags: [dispatch]
      summary: Get dispatch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dispatch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dispatch"
  /dispatch/{id}/approve:
    post:
      operationId: approveDispatch
      tags: [dispatch]
      summary: Approve dispatch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dispatch"
  /dispatch/{id}/dispatch:
    post:
      operationId: startDispatch
      tags: [dispatch]
      summary: Start dispatch (mark as dispatched)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dispatched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dispatch"
  /dispatch/{id}/arrive:
    post:
      operationId: markDispatchArrived
      tags: [dispatch]
      summary: Mark dispatch as arrived
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Arrived
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dispatch"

  # GPS
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]

    # AUTH
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
      required: [token, user]
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        fullName:
          type: string
        email:
          type: string
        role:
          type: string
        districtId:
          type: integer
          nullable: true
        isActive:
          type: boolean
      required: [id, username, fullName, email, role, isActive]

    # DASHBOARD
...
    Campaign:
      type: object
      properties:
        id:
          type: integer
        campaignCode:
          type: string
        name:
          type: string
        season:
          type: string
          nullable: true
        districtId:
          type: integer
        districtName:
          type: string
          nullable: true
        valueChainId:
          type: integer
        valueChainName:
          type: string
          nullable: true
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        distributionSite:
          type: string
          nullable: true
        siteLatitude:
          type: number
          nullable: true
        siteLongitude:
          type: number
          nullable: true
        geofenceRadius:
          type: number
          nullable: true
        status:
          type: string
          enum: [Draft, Submitted, Approved, Active, Completed, Cancelled]
        totalFarmers:
          type: integer
          nullable: true
        totalAllocations:
          type: integer
          nullable: true
        deliveredCount:
          type: integer
          nullable: true
        description:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, campaignCode, name, districtId, valueChainId, startDate, endDate, status, createdAt]
    CreateCampaignBody:
      type: object
      properties:
        name:
          type: string
        season:
          type: string
          nullable: true
        districtId:
          type: integer
        valueChainId:
          type: integer
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        distributionSite:
          type: string
          nullable: true
        siteLatitude:
          type: number
          nullable: true
        siteLongitude:
          type: number
          nullable: true
        geofenceRadius:
          type: number
          nullable: true
        description:
          type: string
          nullable: true
      required: [name, districtId, valueChainId, startDate, endDate]
    UpdateCampaignBody:
      type: object
      properties:
        name:
          type: string
        season:
          type: string
          nullable: true
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        distributionSite:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
    PaginatedCampaigns:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Campaign"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [data, total, page, limit, totalPages]
...
    Allocation:
      type: object
      properties:
        id:
          type: integer
        campaignId:
          type: integer
        campaignName:
          type: string
          nullable: true
        farmerId:
          type: integer
        farmerCode:
          type: string
          nullable: true
        farmerName:
          type: string
          nullable: true
        inputItemId:
          type: integer
        inputItemName:
          type: string
          nullable: true
        quantity:
          type: integer
        status:
          type: string
          enum: [Pending, Allocated, Delivered, Cancelled]
        deliveredAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, campaignId, farmerId, inputItemId, quantity, status, createdAt]
    CreateAllocationBody:
      type: object
      properties:
        campaignId:
          type: integer
        farmerId:
          type: integer
        inputItemId:
          type: integer
        quantity:
          type: integer
      required: [campaignId, farmerId, inputItemId, quantity]
    BulkAllocationBody:
      type: object
      properties:
        campaignId:
          type: integer
        inputItemId:
          type: integer
        quantity:
          type: integer
        farmerIds:
          type: array
          items:
            type: integer
      required: [campaignId, inputItemId, quantity, farmerIds]
    BulkAllocationResult:
      type: object
      properties:
        created:
          type: integer
        skipped:
          type: integer
        errors:
          type: array
          items:
            type: string
      required: [created, skipped, errors]
    PaginatedAllocations:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Allocation"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [data, total, page, limit, totalPages]

    # VEHICLES
...
    Dispatch:
      type: object
      properties:
        id:
          type: integer
        manifestCode:
          type: string
        campaignId:
          type: integer
        campaignName:
          type: string
          nullable: true
        vehicleId:
          type: integer
        vehiclePlate:
          type: string
          nullable: true
        driverId:
          type: integer
          nullable: true
        driverName:
          type: string
          nullable: true
        warehouseId:
          type: integer
        warehouseName:
          type: string
          nullable: true
        status:
          type: string
          enum: [Draft, Approved, Loading, Dispatched, Arrived, Completed, Reconciled, Cancelled]
        totalPackages:
          type: integer
          nullable: true
        deliveredPackages:
          type: integer
          nullable: true
        returnedPackages:
          type: integer
          nullable: true
        dispatchedAt:
          type: string
          format: date-time
          nullable: true
        arrivedAt:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, manifestCode, campaignId, vehicleId, warehouseId, status, createdAt]
    CreateDispatchBody:
      type: object
      properties:
        campaignId:
          type: integer
        vehicleId:
          type: integer
        driverId:
          type: integer
          nullable: true
        warehouseId:
          type: integer
        notes:
          type: string
          nullable: true
      required: [campaignId, vehicleId, warehouseId]
    PaginatedDispatches:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Dispatch"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [data, total, page, limit, totalPages]

    # GPS
...